Consolidate /app/chat and /app/library into /app
Delete standalone library page (redundant with library panel) and move
chat workspace to be the main /app route. Update all navigation refs.
